refactor(cli): extract scheduler + ensure/task/lock to *_cli.py (#495) (#536)
Phase 1 (Wave 2) of #495. Extracts the scheduler command group into
scheduler_cli.py and ensure/task/lock into ensure_cli.py. ensure->new
goes through a deferred import to session_cli. Pure relocation, CLI
surface unchanged. Part of #495.

Built by dotdev.dev
