+++ agentwire/__main__.py
@@ -3693,1866 +3693,6 @@ def _print_tunnel_help(spec, error: str) -> None:
         print(f"        ssh {spec.remote_machine} 'lsof -i :{spec.remote_port}'")
 
 
-# =============================================================================
-# Task Commands (Scheduled Workloads)
-# =============================================================================
-
-# Exit codes for ensure command (documented in CLAUDE.md)
-ENSURE_EXIT_COMPLETE = 0
-ENSURE_EXIT_FAILED = 1
-ENSURE_EXIT_INCOMPLETE = 2
-ENSURE_EXIT_LOCK_CONFLICT = 3
-ENSURE_EXIT_PRE_FAILURE = 4
-ENSURE_EXIT_TIMEOUT = 5
-ENSURE_EXIT_SESSION_ERROR = 6
-ENSURE_EXIT_USAGE_LIMIT = 7
-
-
-def _ensure_remote(args, session: str, machine_id: str, json_mode: bool) -> int:
-    """Delegate `ensure` to the remote machine via SSH.
-
-    When the session target is `name@machine`, we reconstruct the full
-    `agentwire ensure` command and run it on the remote machine natively.
-    All local concerns (locking, idle detection, pre/post commands, summary
-    files) happen on the remote machine where the session actually lives.
-    """
-    import shlex
-
-    machine = _get_machine_config(machine_id)
-    if machine is None:
-        return _output_result(False, json_mode, f"Machine '{machine_id}' not found in machines.json", exit_code=ENSURE_EXIT_SESSION_ERROR)
-
-    host = machine.get("host", machine_id)
-    user = machine.get("user")
-    port = machine.get("port")
-    ssh_target = f"{user}@{host}" if user else host
-
-    # Translate local project path to remote equivalent
-    remote_project = None
-    if hasattr(args, 'project') and args.project:
-        local_path = Path(args.project).expanduser().resolve()
-        # Get local projects dir from config
-        config = load_config()
-        local_projects_dir = Path(config.get("projects", {}).get("dir", "~/projects")).expanduser().resolve()
-        # Get remote projects dir from machine config (or default)
-        remote_projects_dir = machine.get("projects_dir", "~/projects")
-        try:
-            relative = local_path.relative_to(local_projects_dir)
-            remote_project = f"{remote_projects_dir}/{relative}"
-        except ValueError:
-            # Path not under local projects dir — use basename only
-            remote_project = f"{remote_projects_dir}/{local_path.name}"
-
-    # Reconstruct ensure command for the remote (session without @machine)
-    cmd_parts = ["agentwire", "ensure", "-s", session, "--task", args.task, "--json"]
-    if remote_project:
-        cmd_parts.extend(["--project", remote_project])
-    if getattr(args, 'wait_lock', False):
-        cmd_parts.append("--wait-lock")
-    if getattr(args, 'lock_timeout', 60) != 60:
-        cmd_parts.extend(["--lock-timeout", str(args.lock_timeout)])
-    if getattr(args, 'skip_if_locked', False):
-        cmd_parts.append("--skip-if-locked")
-
-    remote_cmd = f"bash -l -c {shlex.quote(' '.join(shlex.quote(p) for p in cmd_parts))}"
-
-    ssh_cmd = ["ssh", "-o", "ConnectTimeout=10", "-o", "BatchMode=yes"]
-    if port:
-        ssh_cmd.extend(["-p", str(port)])
-    ssh_cmd.extend([ssh_target, remote_cmd])
-
-    # Stream output in real-time — ensure can run for tens of minutes
-    try:
-        proc = subprocess.Popen(ssh_cmd, stdout=subprocess.PIPE, stderr=subprocess.STDOUT, text=True)
-        for line in proc.stdout:
-            print(line, end="", flush=True)
-        proc.wait()
-        return proc.returncode
-    except Exception as e:
-        return _output_result(False, json_mode, f"SSH to {machine_id} failed: {e}", exit_code=ENSURE_EXIT_SESSION_ERROR)
-
-
-def cmd_ensure(args) -> int:
-    """Run a named task with reliable session management.
-
-    Full lifecycle:
-    1. Acquire lock (fail if locked, or wait with --wait-lock)
-    2. Ensure session exists and is healthy
-    3. Wait for session to be idle
-    4. Run pre-commands, validate outputs
-    5. Send templated prompt
-    6. Wait for idle, send system summary prompt
-    7. Parse summary file for status
-    8. Send on_task_end prompt if defined
-    9. Run post-commands
-    10. Handle retries on failure
-    """
-    from .completion import (
-        get_summary_prompt,
-    )
-    from .locking import LockConflict, LockTimeout, session_lock
-    from .tasks import (
-        TaskNotFound,
-        TaskValidationError,
-        load_task,
-        validate_task,
-    )
-    from .templating import TemplateContext, preview_template
-
-    session_name = args.session
-    task_name = args.task
-    dry_run = getattr(args, 'dry_run', False)
-    wait_lock = getattr(args, 'wait_lock', False)
-    lock_timeout = getattr(args, 'lock_timeout', 60)
-    skip_if_locked = getattr(args, 'skip_if_locked', False)
-    json_mode = getattr(args, 'json', False)
-
-    # Parse session target
-    session, machine_id = _parse_session_target(session_name)
-
-    if machine_id:
-        return _ensure_remote(args, session, machine_id, json_mode)
-
-    # A session parked on a usage limit is waiting out the reset — never
-    # prompt or re-dispatch into it. The watchdog resumes it.
-    from .usage_limit import is_parked
-    if is_parked(session):
-        return _output_result(
-            False, json_mode,
-            f"Session '{session}' is parked on a usage limit (auto-resumes after reset)",
-            exit_code=ENSURE_EXIT_USAGE_LIMIT,
-        )
-
-    # Find project path from --project flag, or session's working directory
-    if hasattr(args, 'project') and args.project:
-        project_path = Path(args.project).expanduser().resolve()
-    else:
-        project_path = _get_session_project_path(session)
-
-    if not project_path.exists():
-        return _output_result(False, json_mode, f"Project path not found: {project_path}", exit_code=ENSURE_EXIT_SESSION_ERROR)
-
-    # Load task configuration
-    try:
-        task = load_task(project_path, task_name)
-    except TaskNotFound as e:
-        return _output_result(False, json_mode, str(e), exit_code=ENSURE_EXIT_SESSION_ERROR)
-    except TaskValidationError as e:
-        return _output_result(False, json_mode, str(e), exit_code=ENSURE_EXIT_SESSION_ERROR)
-
-    # Validate task
-    issues = validate_task(task)
-    if issues:
-        return _output_result(False, json_mode, f"Task validation failed: {', '.join(issues)}", exit_code=ENSURE_EXIT_SESSION_ERROR)
-
-    # Determine shell
-    shell = task.shell or "/bin/sh"
-
-    # Initialize template context
-    ctx = TemplateContext(
-        session=session,
-        task=task_name,
-        project_root=str(project_path),
-    )
-
-    # Dry run mode
-    if dry_run:
-        print("=== DRY RUN ===\n")
-        print(f"Session: {session}")
-        print(f"Task: {task_name}")
-        print(f"Shell: {shell}")
-        print(f"Idle timeout: {task.idle_timeout}s")
-        print(f"Retries: {task.retries}")
-        print()
-
-        if task.pre:
-            print("Pre-commands (would execute):")
-            for pre in task.pre:
-                req = " (required)" if pre.required else ""
-                val = f" validate: {pre.validate}" if pre.validate else ""
-                print(f"  {pre.name}: {pre.cmd}{req}{val}")
-            print()
-
-        print("Prompt (with placeholders for pre-outputs):")
-        print(preview_template(task.prompt, ctx))
-        print()
-
-        print("System summary prompt:")
-        print(get_summary_prompt("<generated-filename>"))
-        print()
-
-        if task.on_task_end:
-            print("On task end prompt:")
-            print(preview_template(task.on_task_end, ctx))
-            print()
-
-        if task.post:
-            print("Post-commands (would execute):")
-            for cmd in task.post:
-                print(f"  {preview_template(cmd, ctx)}")
-            print()
-
-        if task.output.save:
-            print(f"Save output to: {preview_template(task.output.save, ctx)}")
-
-        return 0
-
-    # Acquire lock
-    try:
-        with session_lock(session, wait=wait_lock, timeout=lock_timeout):
-            return _run_ensure_task(
-                args, session, task, ctx, shell, project_path, json_mode
-            )
-    except LockConflict as e:
-        if skip_if_locked:
-            return 0
-        return _output_result(False, json_mode, str(e), exit_code=ENSURE_EXIT_LOCK_CONFLICT)
-    except LockTimeout as e:
-        if skip_if_locked:
-            return 0
-        return _output_result(False, json_mode, str(e), exit_code=ENSURE_EXIT_LOCK_CONFLICT)
-
-
-def _setup_task_branch(project_path, task, json_mode) -> tuple[str, str | None]:
-    """Set up git branch for a task with starting_ref.
-
-    Checks out starting_ref, pulls latest, creates the work branch.
-
-    Returns:
-        (work_branch_name, error_message) — error_message is None on success.
-    """
-    from .tasks import PreCommandError  # noqa: F401 (used for caller)
-
-    starting_ref = task.starting_ref
-
-    # Verify the ref exists
-    result = subprocess.run(
-        ["git", "rev-parse", "--verify", starting_ref],
-        cwd=project_path,
-        capture_output=True,
-        text=True,
-    )
-    if result.returncode != 0:
-        return "", f"starting_ref '{starting_ref}' not found in {project_path}"
-
-    # Checkout the starting ref
-    checkout = subprocess.run(
-        ["git", "checkout", starting_ref],
-        cwd=project_path,
-        capture_output=True,
-        text=True,
-    )
-    if checkout.returncode != 0:
-        return "", f"Failed to checkout '{starting_ref}': {checkout.stderr.strip()}"
-
-    # Pull if it's a branch (not detached HEAD)
-    head_check = subprocess.run(
-        ["git", "symbolic-ref", "--quiet", "HEAD"],
-        cwd=project_path,
-        capture_output=True,
-    )
-    if head_check.returncode == 0:
-        subprocess.run(
-            ["git", "pull", "--ff-only"],
-            cwd=project_path,
-            capture_output=True,
-        )
-
-    # Determine work branch name
-    work_branch = task.work_branch
-    if not work_branch:
-        today = datetime.date.today().isoformat()
-        work_branch = f"agent/{task.name}-{today}"
-
-    # Handle collision: append -2, -3, ... until name is free
-    check = subprocess.run(
-        ["git", "rev-parse", "--verify", f"refs/heads/{work_branch}"],
-        cwd=project_path,
-        capture_output=True,
-    )
-    if check.returncode == 0:
-        n = 2
-        base = work_branch
-        while True:
-            candidate = f"{base}-{n}"
-            check = subprocess.run(
-                ["git", "rev-parse", "--verify", f"refs/heads/{candidate}"],
-                cwd=project_path,
-                capture_output=True,
-            )
-            if check.returncode != 0:
-                work_branch = candidate
-                break
-            n += 1
-
-    # Create and checkout work branch
-    create = subprocess.run(
-        ["git", "checkout", "-b", work_branch],
-        cwd=project_path,
-        capture_output=True,
-        text=True,
-    )
-    if create.returncode != 0:
-        return "", f"Failed to create work branch '{work_branch}': {create.stderr.strip()}"
-
-    if not json_mode:
-        print(f"Branch: {work_branch} (from {starting_ref})")
-
-    return work_branch, None
-
-
-def _create_task_pr(project_path, task, work_branch, last_summary, json_mode) -> str | None:
-    """Commit, push, and open a PR for completed task work.
-
-    Returns:
-        PR URL if created, None if skipped or failed.
-    """
-    pr_target = task.pr_target or task.starting_ref
-
-    # Check for uncommitted changes
-    status = subprocess.run(
-        ["git", "status", "--porcelain"],
-        cwd=project_path,
-        capture_output=True,
-        text=True,
-    )
-    has_changes = bool(status.stdout.strip())
-
-    if not has_changes:
-        if not json_mode:
-            print("No changes to commit — skipping PR creation")
-        # Still reset to starting_ref
-        subprocess.run(["git", "checkout", task.starting_ref], cwd=project_path, capture_output=True)
-        return None
-
-    # Commit all changes
-    today = datetime.date.today().isoformat()
-    commit_msg = f"chore: agent task {task.name} ({today})"
-    subprocess.run(["git", "add", "-A"], cwd=project_path, capture_output=True)
-    commit_result = subprocess.run(
-        ["git", "commit", "-m", commit_msg],
-        cwd=project_path,
-        capture_output=True,
-        text=True,
-    )
-    if commit_result.returncode != 0:
-        if not json_mode:
-            print(f"Warning: commit failed: {commit_result.stderr.strip()}")
-        subprocess.run(["git", "checkout", task.starting_ref], cwd=project_path, capture_output=True)
-        return None
-
-    # Push branch
-    push = subprocess.run(
-        ["git", "push", "-u", "origin", work_branch],
-        cwd=project_path,
-        capture_output=True,
-        text=True,
-    )
-    if push.returncode != 0:
-        if not json_mode:
-            print(f"Warning: push failed: {push.stderr.strip()}")
-        subprocess.run(["git", "checkout", task.starting_ref], cwd=project_path, capture_output=True)
-        return None
-
-    # Check gh is available
-    gh_check = subprocess.run(["which", "gh"], capture_output=True)
-    if gh_check.returncode != 0:
-        if not json_mode:
-            print("Warning: 'gh' not found — skipping PR creation (branch pushed)")
-        subprocess.run(["git", "checkout", task.starting_ref], cwd=project_path, capture_output=True)
-        return None
-
-    # Create PR
-    pr_title = f"agent: {task.name} ({today})"
-    pr_body = last_summary if last_summary else f"Automated changes from agent task `{task.name}`."
-    pr_cmd = [
-        "gh", "pr", "create",
-        "--base", pr_target,
-        "--head", work_branch,
-        "--title", pr_title,
-        "--body", pr_body,
-    ]
-    if task.pr_draft:
-        pr_cmd.append("--draft")
-
-    pr_result = subprocess.run(
-        pr_cmd,
-        cwd=project_path,
-        capture_output=True,
-        text=True,
-    )
-
-    pr_url = None
-    if pr_result.returncode == 0:
-        pr_url = pr_result.stdout.strip()
-        if not json_mode:
-            print(f"PR created: {pr_url}")
-    else:
-        if not json_mode:
-            print(f"Warning: PR creation failed: {pr_result.stderr.strip()}")
-
-    # Reset to starting_ref
-    subprocess.run(["git", "checkout", task.starting_ref], cwd=project_path, capture_output=True)
-
-    return pr_url
-
-
-def _run_ensure_task(args, session, task, ctx, shell, project_path, json_mode) -> int:
-    """Run the task (called within lock context).
-
-    Uses hook-based completion detection:
-    1. Write task context file (tells hook a scheduled task is running)
-    2. Send task prompt
-    3. Hook handles: first idle → send summary prompt
-    4. Poll for summary file (agent writes it after receiving summary prompt)
-    5. Parse summary and return result
-    """
-    from .completion import (
-        CompletionTimeout,
-        clear_task_context,
-        generate_summary_filename,
-        status_to_exit_code,
-        wait_for_completion_signal,
-        write_task_context,
-    )
-    from .tasks import PreCommandError, run_post_command, run_pre_command
-    from .templating import TemplateError, expand_all
-
-    max_attempts = task.retries + 1
-    last_status = "incomplete"
-    last_summary = ""
-
-    for attempt in range(1, max_attempts + 1):
-        ctx.attempt = attempt
-
-        if not json_mode and max_attempts > 1:
-            print(f"Attempt {attempt}/{max_attempts}")
-
-        # Ensure session exists and has agent running.
-        # The scheduler may have pre-created this session with --model and --type
-        # overrides via _pre_create_session(). Don't kill it — just wait for agent.
-        if not tmux_session_exists(session):
-            if not json_mode:
-                print(f"Creating session '{session}'...")
-
-            # Fork starting_session if configured (carries over Claude conversation context)
-            if task.starting_session and task.starting_session != session:
-                if tmux_session_exists(task.starting_session):
-                    if not json_mode:
-                        print(f"Forking context from session '{task.starting_session}'...")
-                    fork_result = subprocess.run(
-                        ["agentwire", "fork", "-s", task.starting_session, "-t", session, "--json"],
-                        capture_output=True, text=True,
-                    )
-                    if fork_result.returncode != 0 and not json_mode:
-                        print("Warning: context fork failed, starting fresh session")
-                elif not json_mode:
-                    print(f"Warning: starting_session '{task.starting_session}' not found, starting fresh")
-
-            if not tmux_session_exists(session):
-                class NewArgs:
-                    def __init__(self, task_role):
-                        self.session = session
-                        self.path = str(project_path)
-                        self.force = False
-                        self.type = None
-                        self.roles = task_role if task_role else None
-                        self.model = None
-                        self.json = json_mode
-
-                from . import session_cli
-                result = session_cli.cmd_new(NewArgs(task.role))
-                if result != 0:
-                    return _output_result(False, json_mode, f"Failed to create session '{session}'", exit_code=ENSURE_EXIT_SESSION_ERROR)
-
-        # Wait for agent to be ready to accept input.
-        # Handles both freshly-created sessions (agent still loading) and
-        # pre-created sessions from scheduler (agent may be mid-startup).
-        if not json_mode:
-            print("Waiting for agent to be ready...")
-        from agentwire.session_ready import wait_for_session_ready
-        if not wait_for_session_ready(session, timeout=30):
-            # Agent never started — session is dead, bail out
-            if not json_mode:
-                print(f"Agent not ready in session '{session}' after 30s")
-            return _output_result(False, json_mode, f"Agent not running in session '{session}'", exit_code=ENSURE_EXIT_SESSION_ERROR)
-
-        # Set up work branch if starting_ref is configured
-        work_branch = None
-        if task.starting_ref:
-            work_branch, branch_error = _setup_task_branch(project_path, task, json_mode)
-            if branch_error:
-                return _output_result(False, json_mode, branch_error, exit_code=ENSURE_EXIT_PRE_FAILURE)
-
-        # Run pre-commands
-        if task.pre:
-            if not json_mode:
-                print("Running pre-commands...")
-
-            for pre in task.pre:
-                try:
-                    output = run_pre_command(pre, shell, project_path)
-                    ctx.set_pre_output(pre.name, output)
-                    if not json_mode:
-                        print(f"  {pre.name}: {len(output)} chars")
-                except PreCommandError as e:
-                    if work_branch and task.starting_ref:
-                        subprocess.run(["git", "checkout", task.starting_ref], cwd=project_path, capture_output=True)
-                    return _output_result(False, json_mode, str(e), exit_code=ENSURE_EXIT_PRE_FAILURE)
-
-        # Expand prompt
-        try:
-            prompt = expand_all(task.prompt, ctx)
-        except TemplateError as e:
-            return _output_result(False, json_mode, str(e), exit_code=ENSURE_EXIT_PRE_FAILURE)
-
-        # Generate summary filename (scoped to session to avoid collisions)
-        summary_filename = generate_summary_filename(session, task.name)
-        summary_path = project_path / summary_filename
-        ctx.summary_file = summary_filename
-
-        # Ensure .agentwire directory exists
-        (project_path / ".agentwire").mkdir(exist_ok=True)
-
-        # Create iterations directory for loop tasks
-        if task.mode == "loop":
-            (project_path / ".agentwire" / "iterations").mkdir(exist_ok=True)
-
-        # Clear any stale completion signal from a previous run
-        # This prevents immediate return if a previous run's signal wasn't cleaned up
-        clear_task_context(session)
-
-        # Write task context for hook coordination
-        # Hook will: first idle → send summary prompt (ensure polls for summary file directly)
-        # Loop mode: hook iterates (review → re-prompt) until complete or max_iterations
-        write_task_context(
-            session=session,
-            task_name=task.name,
-            summary_file=summary_filename,
-            attempt=attempt,
-            exit_on_complete=task.exit_on_complete,
-            mode=task.mode,
-            max_iterations=task.max_iterations,
-            iteration=1,
-            loop_review=task.loop_review,
-            loop_delay=task.loop_delay,
-            original_prompt=prompt,
-        )
-
-        # Find previous summaries for this task to give the agent context
-        summary_glob = f".agentwire/task-summary-{session}-{task.name}-*.md"
-        prev_summaries = sorted(
-            project_path.glob(summary_glob),
-            key=lambda p: p.stat().st_mtime,
-            reverse=True,
-        )[:5]
-        if prev_summaries:
-            prompt += "\n\nPrevious task summaries (consider them when generating your output):"
-            for p in prev_summaries:
-                prompt += f"\n- {p}"
-
-        if not json_mode:
-            print("Sending task prompt...")
-
-        # Send task prompt using pane_manager for proper multi-line handling
-        pane_manager.send_to_pane(session, 0, prompt, enter=True)
-
-        # Wait for completion signal from hook
-        if not json_mode:
-            print("Waiting for task completion...")
-
-        try:
-            signal = wait_for_completion_signal(
-                session, summary_path=summary_path
-            )
-            last_status = signal.get("status", "incomplete")
-            last_summary = signal.get("summary", "")
-            ctx.status = last_status
-            ctx.summary = last_summary
-        except CompletionTimeout:
-            # Don't clear task context here — the hook may still need it.
-            # Hook cleans up after itself (exit_on_complete kills session).
-            # Task context files are cleared at the START of next run.
-            last_status = "incomplete"
-            last_summary = "Timeout waiting for task completion"
-            if attempt < max_attempts:
-                if not json_mode:
-                    print(f"Timeout, retrying in {task.retry_delay}s...")
-                time.sleep(task.retry_delay)
-                continue
-            break
-
-        # Don't clear task context here — hook owns context file lifecycle.
-        # ensure waits for hook to delete it (signals cleanup complete).
-
-        if last_status == "usage_limit":
-            # Session parked mid-task — skip on_task_end/post/PR; the watchdog
-            # nudges it after reset and the idle hook finishes the task.
-            if not json_mode:
-                print("Usage limit hit — session parked, auto-resumes after reset")
-            break
-
-        if not json_mode:
-            print(f"Task status: {last_status}")
-            if last_summary:
-                print(f"Summary: {last_summary}")
-
-        # on_task_end: send additional prompt after summary is written
-        # Note: we don't wait for this to complete - it's fire-and-forget
-        if task.on_task_end:
-            try:
-                end_prompt = expand_all(task.on_task_end, ctx)
-                pane_manager.send_to_pane(session, 0, end_prompt, enter=True)
-                if not json_mode:
-                    print("Sent on_task_end prompt (not waiting for completion)")
-            except TemplateError as e:
-                if not json_mode:
-                    print(f"Warning: template error in on_task_end: {e}")
-
-        # Capture output
-        output_result = subprocess.run(
-            ["tmux", "capture-pane", "-t", session, "-p", "-S", f"-{task.output.capture}"],
-            capture_output=True,
-            text=True,
-        )
-        ctx.output = output_result.stdout if output_result.returncode == 0 else ""
-
-        # Run post-commands
-        if task.post:
-            if not json_mode:
-                print("Running post-commands...")
-
-            for cmd in task.post:
-                try:
-                    expanded_cmd = expand_all(cmd, ctx)
-                    rc, stdout, stderr = run_post_command(expanded_cmd, shell, project_path)
-                    if rc != 0 and not json_mode:
-                        print(f"  Warning: post-command failed: {stderr}")
-                except TemplateError as e:
-                    if not json_mode:
-                        print(f"  Warning: template error in post-command: {e}")
-
-        # Save output if configured
-        if task.output.save:
-            try:
-                save_path = Path(expand_all(task.output.save, ctx)).expanduser()
-                save_path.parent.mkdir(parents=True, exist_ok=True)
-                save_path.write_text(ctx.output)
-                if not json_mode:
-                    print(f"Output saved to: {save_path}")
-            except Exception as e:
-                if not json_mode:
-                    print(f"Warning: Failed to save output: {e}")
-
-        # Create PR if branch management is configured
-        if work_branch and task.starting_ref:
-            pr_url = _create_task_pr(project_path, task, work_branch, last_summary, json_mode)
-            ctx.work_branch = work_branch
-            if pr_url:
-                ctx.pr_url = pr_url
-
-        # Check if we should retry
-        if last_status == "failed" and attempt < max_attempts:
-            if not json_mode:
-                print(f"Task failed, retrying in {task.retry_delay}s...")
-            time.sleep(task.retry_delay)
-            continue
-
-        # Done (success or no more retries)
-        break
-
-    # Final result
-    exit_code = status_to_exit_code(last_status)
-
-    if json_mode:
-        result_data = {
-            "success": last_status == "complete",
-            "status": last_status,
-            "summary": last_summary,
-            "attempt": ctx.attempt,
-            "summary_file": ctx.summary_file,
-        }
-        if ctx.work_branch:
-            result_data["work_branch"] = ctx.work_branch
-        if ctx.pr_url:
-            result_data["pr_url"] = ctx.pr_url
-        _output_json(result_data)
-    else:
-        print(f"\nTask {task.name}: {last_status}")
-
-    return exit_code
-
-
-def cmd_task_list(args) -> int:
-    """List tasks for a session/project."""
-    from .tasks import list_tasks
-
-    session = getattr(args, 'session', None)
-    json_mode = getattr(args, 'json', False)
-
-    # Find project path from session's working directory or cwd
-    if session:
-        project_path = _get_session_project_path(session)
-    else:
-        project_path = Path.cwd()
-
-    if not project_path or not project_path.exists():
-        return _output_result(False, json_mode, f"Project path not found: {project_path}")
-
-    tasks = list_tasks(project_path)
-
-    if json_mode:
-        _output_json({"tasks": tasks, "project": str(project_path)})
-        return 0
-
-    if not tasks:
-        print(f"No tasks defined in {project_path / '.agentwire.yml'}")
-        return 0
-
-    print(f"Tasks in {project_path.name}:\n")
-    print(f"{'Name':<25} {'Mode':<10} {'Pre':<5} {'Post':<5} {'Retries':<8}")
-    print("-" * 60)
-    for t in tasks:
-        pre = "Yes" if t["has_pre"] else "-"
-        post = "Yes" if t["has_post"] else "-"
-        mode = t.get("mode", "standard")
-        print(f"{t['name']:<25} {mode:<10} {pre:<5} {post:<5} {t['retries']:<8}")
-
-    return 0
-
-
-def cmd_task_show(args) -> int:
-    """Show task definition details."""
-    from .tasks import TaskNotFound, TaskValidationError, load_task, validate_task
-
-    task_arg = args.task  # format: session/task or just task
-    json_mode = getattr(args, 'json', False)
-
-    # Parse task argument
-    if "/" in task_arg:
-        session, task_name = task_arg.split("/", 1)
-    else:
-        session = None
-        task_name = task_arg
-
-    # Find project path from session's working directory or cwd
-    if session:
-        project_path = _get_session_project_path(session)
-    else:
-        project_path = Path.cwd()
-
-    try:
-        task = load_task(project_path, task_name)
-    except (TaskNotFound, TaskValidationError) as e:
-        return _output_result(False, json_mode, str(e))
-
-    issues = validate_task(task)
-
-    if json_mode:
-        _output_json({
-            "name": task.name,
-            "prompt": task.prompt,
-            "shell": task.shell,
-            "retries": task.retries,
-            "retry_delay": task.retry_delay,
-            "idle_timeout": task.idle_timeout,
-            "mode": task.mode,
-            "max_iterations": task.max_iterations,
-            "loop_review": task.loop_review,
-            "loop_delay": task.loop_delay,
-            "pre": [{"name": p.name, "cmd": p.cmd, "required": p.required, "validate": p.validate, "timeout": p.timeout} for p in task.pre],
-            "on_task_end": task.on_task_end,
-            "post": task.post,
-            "output": {"capture": task.output.capture, "save": task.output.save},
-            "validation_issues": issues,
-        })
-        return 0
-
-    print(f"Task: {task.name}\n")
-    print(f"Shell: {task.shell or '/bin/sh'}")
-    print(f"Mode: {task.mode}")
-    if task.mode == "loop":
-        print(f"Max iterations: {task.max_iterations}")
-        print(f"Loop review: {task.loop_review}")
-        if task.loop_delay > 0:
-            print(f"Loop delay: {task.loop_delay}s")
-    print(f"Retries: {task.retries} (delay: {task.retry_delay}s)")
-    print(f"Idle timeout: {task.idle_timeout}s")
-    print()
-
-    if task.pre:
-        print("Pre-commands:")
-        for p in task.pre:
-            req = " (required)" if p.required else ""
-            print(f"  {p.name}: {p.cmd}{req}")
-        print()
-
-    print("Prompt:")
-    print(task.prompt[:200] + "..." if len(task.prompt) > 200 else task.prompt)
-    print()
-
-    if task.on_task_end:
-        print("On task end:")
-        print(task.on_task_end[:100] + "..." if len(task.on_task_end) > 100 else task.on_task_end)
-        print()
-
-    if task.post:
-        print("Post-commands:")
-        for cmd in task.post:
-            print(f"  {cmd}")
-        print()
-
-    if task.output.save:
-        print("Output:")
-        print(f"  Save to: {task.output.save}")
-
-    if issues:
-        print(f"\nValidation issues: {', '.join(issues)}")
-
-    return 0
-
-
-def cmd_task_validate(args) -> int:
-    """Validate task configuration."""
-    from .tasks import TaskNotFound, TaskValidationError, load_task, validate_task
-
-    task_arg = args.task
-    json_mode = getattr(args, 'json', False)
-
-    # Parse task argument
-    if "/" in task_arg:
-        session, task_name = task_arg.split("/", 1)
-    else:
-        session = None
-        task_name = task_arg
-
-    # Find project path from session's working directory or cwd
-    if session:
-        project_path = _get_session_project_path(session)
-    else:
-        project_path = Path.cwd()
-
-    try:
-        task = load_task(project_path, task_name)
-    except (TaskNotFound, TaskValidationError) as e:
-        return _output_result(False, json_mode, str(e))
-
-    issues = validate_task(task)
-
-    if json_mode:
-        _output_json({
-            "valid": len(issues) == 0,
-            "issues": issues,
-            "task": task_name,
-        })
-        return 0 if not issues else 1
-
-    if issues:
-        print(f"Task '{task_name}' has issues:")
-        for issue in issues:
-            print(f"  - {issue}")
-        return 1
-    else:
-        print(f"Task '{task_name}' is valid.")
-        return 0
-
-
-# =============================================================================
-# Lock Management Commands
-# =============================================================================
-
-
-def cmd_lock_list(args) -> int:
-    """List all locks with metadata."""
-    from .locking import list_locks
-
-    json_mode = getattr(args, 'json', False)
-    locks = list_locks()
-
-    if json_mode:
-        _output_json({"locks": locks})
-        return 0
-
-    if not locks:
-        print("No locks found.")
-        return 0
-
-    # Format output
-    print(f"{'SESSION':<25} {'PID':<10} {'AGE':<12} {'STATUS'}")
-    print("-" * 60)
-
-    for lock in locks:
-        session = lock["session"][:24]
-        pid = str(lock["pid"]) if lock["pid"] else "-"
-        age_seconds = lock["age_seconds"]
-
-        # Format age
-        if age_seconds < 60:
-            age = f"{age_seconds}s"
-        elif age_seconds < 3600:
-            age = f"{age_seconds // 60}m {age_seconds % 60}s"
-        elif age_seconds < 86400:
-            hours = age_seconds // 3600
-            mins = (age_seconds % 3600) // 60
-            age = f"{hours}h {mins}m"
-        else:
-            days = age_seconds // 86400
-            hours = (age_seconds % 86400) // 3600
-            age = f"{days}d {hours}h"
-
-        status = lock["status"]
-        print(f"{session:<25} {pid:<10} {age:<12} {status}")
-
-    return 0
-
-
-def cmd_lock_clean(args) -> int:
-    """Remove all stale locks."""
-    from .locking import clean_stale_locks
-
-    json_mode = getattr(args, 'json', False)
-    dry_run = getattr(args, 'dry_run', False)
-
-    removed = clean_stale_locks(dry_run=dry_run)
-
-    if json_mode:
-        _output_json({
-            "removed": removed,
-            "count": len(removed),
-            "dry_run": dry_run,
-        })
-        return 0
-
-    if not removed:
-        print("No stale locks found.")
-    elif dry_run:
-        print(f"Would remove {len(removed)} stale lock(s): {', '.join(removed)}")
-    else:
-        print(f"Removed {len(removed)} stale lock(s): {', '.join(removed)}")
-
-    return 0
-
-
-def cmd_lock_remove(args) -> int:
-    """Force-remove a specific lock."""
-    from .locking import remove_lock
-
-    session = args.session
-    json_mode = getattr(args, 'json', False)
-
-    removed = remove_lock(session)
-
-    if json_mode:
-        _output_json({
-            "session": session,
-            "removed": removed,
-        })
-        return 0 if removed else 1
-
-    if removed:
-        print(f"Removed lock: {session}")
-        return 0
-    else:
-        print(f"No lock found for: {session}")
-        return 1
-
-
-def cmd_lock(args) -> int:
-    """Lock command dispatcher - shows help if no subcommand."""
-    # This will be called if no subcommand is provided
-    # The help is printed in main() based on lock_command being None
-    return 0
-
-
-# =============================================================================
-# Scheduler Commands
-# =============================================================================
-
-
-SCHEDULER_SESSION = "agentwire-scheduler"
-
-
-def cmd_scheduler_start(args) -> int:
-    """Start the scheduler daemon in a tmux session."""
-    if not _check_tmux_installed():
-        return 1
-
-    if tmux_session_exists(SCHEDULER_SESSION):
-        print(f"Scheduler already running in tmux session '{SCHEDULER_SESSION}'")
-        print("Attaching... (Ctrl+B D to detach)")
-        subprocess.run(["tmux", "attach-session", "-t", SCHEDULER_SESSION])
-        return 0
-
-    print(f"Starting scheduler daemon in tmux session '{SCHEDULER_SESSION}'...")
-    subprocess.run([
-        "tmux", "new-session", "-d", "-s", SCHEDULER_SESSION,
-    ])
-    subprocess.run([
-        "tmux", "send-keys", "-t", SCHEDULER_SESSION,
-        "agentwire scheduler serve", "Enter",
-    ])
-
-    print("Attaching... (Ctrl+B D to detach)")
-    subprocess.run(["tmux", "attach-session", "-t", SCHEDULER_SESSION])
-    return 0
-
-
-def cmd_scheduler_serve(args) -> int:
-    """Run the scheduler loop in the foreground (for tmux)."""
-    from .scheduler import run_scheduler_loop
-
-    run_scheduler_loop()
-    return 0
-
-
-def cmd_scheduler_stop(args) -> int:
-    """Stop the scheduler daemon."""
-    if not tmux_session_exists(SCHEDULER_SESSION):
-        print("Scheduler is not running.")
-        return 1
-
-    subprocess.run(["tmux", "kill-session", "-t", SCHEDULER_SESSION])
-    print("Scheduler stopped.")
-    return 0
-
-
-def cmd_scheduler_status(args) -> int:
-    """Show scheduler status and next task due."""
-    from .config import get_config
-    from .scheduler import (
-        format_interval,
-        load_board,
-        pick_next_task,
-        read_events,
-    )
-
-    json_mode = getattr(args, 'json', False)
-    running = tmux_session_exists(SCHEDULER_SESSION)
-    board_path = get_config().scheduler.board_file
-
-    if not board_path.exists():
-        return _output_result(
-            False, json_mode,
-            f"Board file not found: {board_path}",
-            running=running,
-        )
-
-    try:
-        board = load_board()
-    except (FileNotFoundError, ValueError) as e:
-        return _output_result(False, json_mode, str(e), running=running)
-
-    task_count = len(board.tasks)
-    enabled_count = sum(1 for t in board.tasks.values() if t.enabled)
-    next_task, wait_seconds = pick_next_task(board)
-    recent_activity = _recent_activity(read_events(tail=60), limit=5)
-
-    result = {
-        "running": running,
-        "board_path": str(board_path),
-        "task_count": task_count,
-        "enabled_count": enabled_count,
-        "next_task": next_task,
-        "next_in_seconds": round(wait_seconds, 1),
-        "recent_activity": recent_activity,
-    }
-
-    if json_mode:
-        _output_json({"success": True, **result})
-        return 0
-
-    status_str = "running" if running else "stopped"
-    print(f"Scheduler: {status_str}")
-    print(f"Board: {board_path}")
-    print(f"Tasks: {enabled_count}/{task_count} enabled")
-
-    if next_task:
-        if wait_seconds <= 0:
-            print(f"Next: {next_task} (due now)")
-        else:
-            print(f"Next: {next_task} (in {format_interval(int(wait_seconds))})")
-    else:
-        print("Next: nothing due")
-
-    if recent_activity:
-        print("\nRecent activity:")
-        for item in recent_activity:
-            print(f"  {item['when']:<16} {item['task']:<24} {item['detail']}")
-
-    return 0
-
-
-def _recent_activity(events: list[dict], limit: int = 5) -> list[dict]:
-    """Distill the event stream into a short 'what just happened' list.
-
-    Keeps the outcome-bearing events (completed/failed/gate-error) so a
-    glance at `scheduler status` shows recent results — including the
-    fail-open gate errors that used to vanish entirely.
-    """
-    keep = {"task_completed", "task_failed", "gate_error"}
-    out: list[dict] = []
-    for evt in reversed(events):
-        etype = evt.get("event")
-        if etype not in keep:
-            continue
-        ts = evt.get("ts", "")
-        try:
-            when = datetime.datetime.fromisoformat(ts).strftime("%m-%d %H:%M")
-        except (ValueError, TypeError):
-            when = ts[:16] if ts else "?"
-        if etype == "task_completed":
-            status = evt.get("status", "?")
-            summary = evt.get("summary", "")
-            detail = f"{status}" + (f" — {summary}" if summary else "")
-        elif etype == "task_failed":
-            detail = "failed — " + (evt.get("summary") or evt.get("reason") or "?")
-        else:  # gate_error
-            detail = f"[gate-error] {evt.get('gate_type', '?')}: {evt.get('reason', '?')}"
-        if len(detail) > 80:
-            detail = detail[:79] + "…"
-        out.append({"when": when, "task": evt.get("task", "?"), "detail": detail})
-        if len(out) >= limit:
-            break
-    return out
-
-
-# Statuses whose last_summary is worth surfacing as a "why" line on the board.
-_BAD_STATUSES = {"failed", "incomplete", "timeout", "lock_conflict", "usage_limit"}
-
-
-def cmd_scheduler_board(args) -> int:
-    """Show full task board with overdue scores."""
-    from .scheduler import get_board_display, load_board
-
-    json_mode = getattr(args, 'json', False)
-
-    try:
-        board = load_board()
-    except (FileNotFoundError, ValueError) as e:
-        return _output_result(False, json_mode, str(e))
-
-    rows = get_board_display(board)
-
-    if json_mode:
-        _output_json({"success": True, "tasks": rows})
-        return 0
-
-    if not rows:
-        print("No tasks in board.")
-        return 0
-
-    # Group by project
-    groups: dict[str, list[dict]] = {}
-    for r in rows:
-        proj = r["project"].rstrip("/").split("/")[-1]
-        groups.setdefault(proj, [])
-        groups[proj].append(r)
-
-    # Summary line
-    total = len(rows)
-    regular = sum(1 for r in rows if not r["filler"])
-    fillers = total - regular
-    enabled = sum(1 for r in rows if r["enabled"])
-    print(f"Scheduler board: {total} tasks ({regular} regular + {fillers} filler), {enabled} enabled\n")
-
-    for proj, items in groups.items():
-        # Sort: regular first (by priority), then filler (by priority)
-        reg = sorted([r for r in items if not r["filler"]], key=lambda r: r["priority"])
-        fil = sorted([r for r in items if r["filler"]], key=lambda r: r["priority"])
-
-        session = items[0]["session"]
-        print(f"  {proj} ({len(items)} tasks) → {session}")
-        print(f"  {'Task':<30} {'Type':<16} {'Schedule':<24} {'Last Run':<16} {'Status':<12} {'Overdue'}")
-        print(f"  {'-' * 114}")
-
-        for r in reg + fil:
-            task_name = r["task"]
-            if not r["enabled"]:
-                task_name = f"{task_name} [off]"
-
-            if r["filler"]:
-                type_str = f"filler (p{r['priority']})"
-            elif r["priority"] != 99:
-                type_str = f"regular (p{r['priority']})"
-            else:
-                type_str = "regular"
-
-            status_str = r["last_status"]
-            if r.get("in_flight"):
-                status_str = "[in-flight]"
-
-            schedule_display = r.get("schedule_str", "?")
-            if len(schedule_display) > 22:
-                schedule_display = schedule_display[:21] + "…"
-
-            print(
-                f"  {task_name:<30} "
-                f"{type_str:<16} "
-                f"{schedule_display:<24} "
-                f"{r['last_run']:<16} "
-                f"{status_str:<12} "
-                f"{r['overdue_str']}"
-            )
-
-            # Surface WHY: a gate-eval error (fail-open, would otherwise be
-            # invisible) takes precedence; else the summary behind a bad status.
-            detail = ""
-            if r.get("last_gate_error"):
-                detail = f"[gate-error] {r['last_gate_error']}"
-            elif status_str in _BAD_STATUSES and r.get("last_summary"):
-                detail = r["last_summary"]
-            if detail:
-                if len(detail) > 96:
-                    detail = detail[:95] + "…"
-                print(f"  {'':<30} ↳ {detail}")
-
-        print()
-
-    return 0
-
-
-def cmd_scheduler_run(args) -> int:
-    """Force-run a specific task now."""
-    from .scheduler import dispatch_task, load_board, save_board
-
-    json_mode = getattr(args, 'json', False)
-    name = args.name
-
-    try:
-        board = load_board()
-    except (FileNotFoundError, ValueError) as e:
-        return _output_result(False, json_mode, str(e))
-
-    if name not in board.tasks:
-        return _output_result(
-            False, json_mode,
-            f"Task '{name}' not found in board. Available: {', '.join(board.tasks.keys())}",
-        )
-
-    if not json_mode:
-        print(f"Running: {name}")
-
-    state = dispatch_task(board, name)
-    board.state[name] = state
-    save_board(board)
-
-    if json_mode:
-        _output_json({
-            "success": state.last_status == "complete",
-            "task": name,
-            "status": state.last_status,
-            "duration": state.last_duration,
-            "run_count": state.run_count,
-        })
-        return 0 if state.last_status == "complete" else 1
-
-    print(f"Done: {name} → {state.last_status} ({state.last_duration}s)")
-    return 0 if state.last_status == "complete" else 1
-
-
-def cmd_scheduler_enable(args) -> int:
-    """Enable a task in the board."""
-    return _set_task_enabled(args.name, True)
-
-
-def cmd_scheduler_disable(args) -> int:
-    """Disable a task in the board."""
-    return _set_task_enabled(args.name, False)
-
-
-def _set_task_enabled(name: str, enabled: bool) -> int:
-    """Toggle a task's enabled field in the board YAML."""
-    import yaml
-
-    from .config import get_config
-
-    board_path = get_config().scheduler.board_file
-
-    if not board_path.exists():
-        print(f"Board file not found: {board_path}", file=sys.stderr)
-        return 1
-
-    with open(board_path) as f:
-        raw = yaml.safe_load(f) or {}
-
-    tasks = raw.get("tasks", {})
-    if name not in tasks:
-        print(f"Task '{name}' not found in board.", file=sys.stderr)
-        return 1
-
-    tasks[name]["enabled"] = enabled
-
-    # Atomic + validated write — never leave scheduler.yaml half-written (#449).
-    from .scheduler import _atomic_write
-
-    text = yaml.dump(raw, default_flow_style=False, sort_keys=False)
-
-    def _validate(tmp_path: str) -> None:
-        with open(tmp_path) as f:
-            reparsed = yaml.safe_load(f)
-        if not isinstance(reparsed, dict) or "tasks" not in reparsed:
-            raise ValueError("scheduler board failed re-parse validation")
-
-    _atomic_write(board_path, text, validate=_validate)
-
-    action = "Enabled" if enabled else "Disabled"
-    print(f"{action}: {name}")
-    return 0
-
-
-def cmd_scheduler_history(args) -> int:
-    """Show recent run history from board state."""
-    from .scheduler import format_interval, load_board
-
-    json_mode = getattr(args, 'json', False)
-
-    try:
-        board = load_board()
-    except (FileNotFoundError, ValueError) as e:
-        return _output_result(False, json_mode, str(e))
-
-    if json_mode:
-        history = []
-        for name, state in board.state.items():
-            history.append({
-                "task": name,
-                "last_run": state.last_run.isoformat() if state.last_run else None,
-                "last_status": state.last_status,
-                "last_duration": state.last_duration,
-                "run_count": state.run_count,
-            })
-        _output_json({"success": True, "history": history})
-        return 0
-
-    if not board.state:
-        print("No run history.")
-        return 0
-
-    print(f"{'Task':<30} {'Last Run':<20} {'Status':<14} {'Duration':<10} {'Runs'}")
-    print("-" * 85)
-
-    for name, state in sorted(board.state.items()):
-        if state.last_run:
-            lr = state.last_run.strftime("%Y-%m-%d %H:%M")
-        else:
-            lr = "never"
-
-        dur = format_interval(state.last_duration) if state.last_duration else "-"
-        print(f"{name:<30} {lr:<20} {state.last_status:<14} {dur:<10} {state.run_count}")
-
-    return 0
-
-
-def cmd_scheduler_report(args) -> int:
-    """Generate a morning report HTML artifact of recent task runs."""
-
-    from .scheduler import _parse_duration, format_interval, load_board, read_events
-
-    json_mode = getattr(args, 'json', False)
-    since_str = getattr(args, 'since', '8h') or '8h'
-    open_artifact = getattr(args, 'artifact', False)
-
-    # Parse duration
-    since_seconds = _parse_duration(since_str) or 28800  # default 8h
-    cutoff = datetime.datetime.now(datetime.timezone.utc) - datetime.timedelta(seconds=since_seconds)
-
-    # Load board state (validate it loads; events are read below)
-    try:
-        load_board()
-    except Exception as e:
-        print(f"Error loading board: {e}", file=sys.stderr)
-        return 1
-
-    # Load events in the window
-    try:
-        events = read_events(tail=500)
-    except Exception:
-        events = []
-
-    # Collect completed task events within window
-    runs: list[dict] = []
-    for ev in events:
-        if ev.get("event") != "task_completed":
-            continue
-        ts_str = ev.get("ts") or ev.get("timestamp", "")
-        try:
-            ts = datetime.datetime.fromisoformat(ts_str.replace("Z", "+00:00"))
-        except (ValueError, AttributeError):
-            continue
-        if ts < cutoff:
-            continue
-        task_name = ev.get("task", "")
-        # Collect run data
-        run = {
-            "task": task_name,
-            "status": ev.get("status", "unknown"),
-            "duration": ev.get("duration", 0),
-            "summary": ev.get("summary", ""),
-            "timestamp": ts.strftime("%Y-%m-%d %H:%M"),
-            "work_branch": "",
-            "pr_url": "",
-            "workflow": ev.get("workflow", ""),
-            "run_id": ev.get("run_id", ""),
-            "nodes": ev.get("nodes") or [],
-        }
-        runs.append(run)
-
-    # Count totals
-    total = len(runs)
-    complete = sum(1 for r in runs if r["status"] == "complete")
-    failed = sum(1 for r in runs if r["status"] in ("failed", "timeout"))
-    incomplete = total - complete - failed
-
-    now_str = datetime.datetime.now().strftime("%Y-%m-%d %H:%M")
-    report_date = datetime.datetime.now().strftime("%Y-%m-%d")
-
-    def status_badge(status: str) -> str:
-        colors = {
-            "complete": "#00c853",
-            "failed": "#ff5252",
-            "timeout": "#ff7043",
-            "incomplete": "#ffa726",
-            "unknown": "#78909c",
-        }
-        color = colors.get(status, "#78909c")
-        return f'<span style="background:{color};color:#fff;padding:2px 8px;border-radius:12px;font-size:0.85em">{status}</span>'
-
-    rows_html = ""
-    for r in runs:
-        duration_str = format_interval(r["duration"]) if r["duration"] else "-"
-        pr_link = f'<a href="{r["pr_url"]}" target="_blank" style="color:#00d4ff">{r["pr_url"][:40]}...</a>' if r.get("pr_url") else "-"
-        branch_col = f'<code style="font-size:0.85em">{r.get("work_branch") or "-"}</code>'
-        summary_text = r["summary"][:120] if r["summary"] else "-"
-        rows_html += f"""
-        <tr>
-          <td style="font-weight:600">{r["task"]}</td>
-          <td>{status_badge(r["status"])}</td>
-          <td>{r["timestamp"]}</td>
-          <td>{duration_str}</td>
-          <td>{branch_col}</td>
-          <td>{pr_link}</td>
-          <td style="color:#aaa;font-size:0.85em">{summary_text}</td>
-        </tr>"""
-
-    if not rows_html:
-        rows_html = f'<tr><td colspan="7" style="color:#556;text-align:center;padding:24px">No tasks ran in the last {since_str}</td></tr>'
-
-    html = f"""<!DOCTYPE html>
-<html>
-<head>
-<meta charset="utf-8">
-<title>Morning Report — {report_date}</title>
-<style>
-  * {{ box-sizing: border-box; margin: 0; padding: 0; }}
-  body {{ font-family: -apple-system, BlinkMacSystemFont, 'Segoe UI', sans-serif; margin: 24px; background: #1a1a2e; color: #e0e0e0; }}
-  h1 {{ color: #00d4ff; margin-bottom: 4px; }}
-  .meta {{ color: #556; font-size: 0.85em; margin-bottom: 20px; }}
-  .summary-bar {{ display: flex; gap: 24px; padding: 14px 20px; background: #16213e; border-radius: 8px; margin-bottom: 24px; }}
-  .summary-bar .item {{ display: flex; flex-direction: column; }}
-  .summary-bar .label {{ color: #556; font-size: 0.75em; text-transform: uppercase; letter-spacing: 0.5px; }}
-  .summary-bar .value {{ font-size: 1.4em; font-weight: 700; }}
-  .complete {{ color: #00c853; }}
-  .failed {{ color: #ff5252; }}
-  .incomplete {{ color: #ffa726; }}
-  .total {{ color: #e0e0e0; }}
-  table {{ border-collapse: collapse; width: 100%; }}
-  th, td {{ text-align: left; padding: 8px 12px; border-bottom: 1px solid #2a2a4a; font-size: 0.9em; }}
-  th {{ background: #16213e; color: #00d4ff; font-weight: 600; position: sticky; top: 0; }}
-  tr:hover {{ background: #16213e; }}
-  code {{ background: #0d1b2a; padding: 2px 6px; border-radius: 4px; }}
-</style>
-</head>
-<body>
-<h1>Morning Report</h1>
-<p class="meta">Generated {now_str} &nbsp;&middot;&nbsp; Last {since_str}</p>
-
-<div class="summary-bar">
-  <div class="item"><span class="label">Total</span><span class="value total">{total}</span></div>
-  <div class="item"><span class="label">Complete</span><span class="value complete">{complete}</span></div>
-  <div class="item"><span class="label">Failed</span><span class="value failed">{failed}</span></div>
-  <div class="item"><span class="label">Incomplete</span><span class="value incomplete">{incomplete}</span></div>
-</div>
-
-<table>
-  <thead>
-    <tr>
-      <th>Task</th>
-      <th>Status</th>
-      <th>Time</th>
-      <th>Duration</th>
-      <th>Branch</th>
-      <th>PR</th>
-      <th>Summary</th>
-    </tr>
-  </thead>
-  <tbody>{rows_html}</tbody>
-</table>
-</body>
-</html>"""
-
-    # Write artifact
-    artifacts_dir = Path.home() / ".agentwire" / "artifacts"
-    artifacts_dir.mkdir(parents=True, exist_ok=True)
-    filename = f"morning-report-{report_date}.html"
-    report_path = artifacts_dir / filename
-    report_path.write_text(html)
-
-    if json_mode:
-        _output_json({
-            "success": True,
-            "path": str(report_path),
-            "filename": filename,
-            "total": total,
-            "complete": complete,
-            "failed": failed,
-            "incomplete": incomplete,
-        })
-    else:
-        print(f"Report: {report_path}")
-        print(f"Tasks: {total} total — {complete} complete, {failed} failed, {incomplete} incomplete")
-
-    if open_artifact:
-        subprocess.run(
-            ["agentwire", "open", filename, "--title", f"Morning Report {report_date}"],
-            capture_output=True,
-        )
-
-    return 0
-
-
-def cmd_scheduler_events(args) -> int:
-    """Show recent scheduler events from the JSONL log."""
-    from .scheduler import read_events
-
-    json_mode = getattr(args, 'json', False)
-    tail = getattr(args, 'tail', 20)
-    task_filter = getattr(args, 'task', None)
-
-    events = read_events(tail=tail, task_filter=task_filter)
-
-    if json_mode:
-        _output_json({"success": True, "events": events})
-        return 0
-
-    if not events:
-        print("No scheduler events.")
-        return 0
-
-    for evt in events:
-        ts = evt.get("ts", "")
-        # Format timestamp for display
-        try:
-            dt = datetime.datetime.fromisoformat(ts)
-            ts_str = dt.strftime("%Y-%m-%d %H:%M")
-        except (ValueError, TypeError):
-            ts_str = ts[:16] if ts else "?"
-
-        event_type = evt.get("event", "?")
-        task_name = evt.get("task", "")
-        session = evt.get("session", "")
-
-        if event_type == "task_completed":
-            status = evt.get("status", "?")
-            duration = evt.get("duration", 0)
-            summary = evt.get("summary", "")
-            summary_str = f'  "{summary}"' if summary else ""
-            print(f"{ts_str}  {event_type:<22} {task_name:<24} {status:<12} {duration}s{summary_str}")
-        elif event_type == "task_started":
-            print(f"{ts_str}  {event_type:<22} {task_name:<24} {session}")
-        elif event_type == "task_skipped":
-            reason = evt.get("reason", "?")
-            print(f"{ts_str}  {event_type:<22} {task_name:<24} reason: {reason}")
-        elif event_type == "gate_error":
-            gate_type = evt.get("gate_type", "?")
-            reason = evt.get("reason", "?")
-            print(f"{ts_str}  {event_type:<22} {task_name:<24} {gate_type}: {reason} (failed open)")
-        elif event_type == "scheduler_sleeping":
-            next_task = evt.get("next_task", "?")
-            sleep_s = evt.get("sleep_seconds", 0)
-            print(f"{ts_str}  {event_type:<22} next: {next_task} in {int(sleep_s)}s")
-        elif event_type == "scheduler_started":
-            count = evt.get("task_count", 0)
-            enabled = evt.get("enabled_count", 0)
-            print(f"{ts_str}  {event_type:<22} {enabled}/{count} tasks enabled")
-        else:
-            print(f"{ts_str}  {event_type:<22} {task_name}")
-
-    return 0
-
-
-def cmd_scheduler_live(args) -> int:
-    """Show live scheduler state."""
-    from .scheduler import format_interval, read_live_state
-
-    json_mode = getattr(args, 'json', False)
-    watch_mode = getattr(args, 'watch', False)
-
-    def _display_once():
-        state = read_live_state()
-        if not state:
-            if json_mode:
-                _output_json({"success": False, "error": "No live state file. Is the scheduler running?"})
-            else:
-                print("No live state available. Is the scheduler running?")
-            return False
-
-        if json_mode:
-            _output_json({"success": True, **state})
-            return True
-
-        status = state.get("status", "unknown")
-        uptime = state.get("uptime_seconds", 0)
-        current = state.get("current_task")
-        current_started = state.get("current_task_started")
-        completed = state.get("tasks_completed", 0)
-        failed = state.get("tasks_failed", 0)
-        next_task = state.get("next_task")
-        next_in = state.get("next_in_seconds", 0)
-
-        print(f"Scheduler: {status} (uptime {format_interval(int(uptime))})")
-
-        if current:
-            # Calculate running time
-            running_str = ""
-            if current_started:
-                try:
-                    started_dt = datetime.datetime.fromisoformat(current_started)
-                    running = int((datetime.datetime.now(datetime.timezone.utc) - started_dt).total_seconds())
-                    running_str = f" (running {format_interval(running)})"
-                except (ValueError, TypeError):
-                    pass
-            print(f"Current:   {current}{running_str}")
-        else:
-            print("Current:   idle")
-
-        print(f"Completed: {completed} tasks | Failed: {failed}")
-
-        if next_task:
-            print(f"Next:      {next_task} (in {format_interval(int(next_in))})")
-        elif not current:
-            print("Next:      nothing due")
-
-        return True
-
-    if watch_mode:
-        import os
-        try:
-            while True:
-                os.system("clear")
-                _display_once()
-                time.sleep(2)
-        except KeyboardInterrupt:
-            return 0
-    else:
-        _display_once()
-        return 0
-
-
-def cmd_scheduler_dashboard(args) -> int:
-    """Generate and open a live HTML dashboard as a portal artifact."""
-    no_open = getattr(args, 'no_open', False)
-
-    html = _generate_dashboard_html()
-
-    # Write to artifacts
-    artifacts_dir = Path.home() / ".agentwire" / "artifacts"
-    artifacts_dir.mkdir(parents=True, exist_ok=True)
-    dashboard_path = artifacts_dir / "scheduler-dashboard.html"
-    dashboard_path.write_text(html)
-    print(f"Dashboard written to {dashboard_path}")
-
-    if not no_open:
-        subprocess.run(
-            ["agentwire", "open", "scheduler-dashboard.html", "--title", "Scheduler Dashboard"],
-            capture_output=True,
-        )
-
-    return 0
-
-
-def _generate_dashboard_html() -> str:
-    """Generate a live scheduler dashboard that fetches data from REST APIs."""
-
-    return '''<!DOCTYPE html>
-<html>
-<head>
-<meta charset="utf-8">
-<title>Scheduler Dashboard</title>
-<style>
-  * { box-sizing: border-box; margin: 0; padding: 0; }
-  body { font-family: -apple-system, BlinkMacSystemFont, 'Segoe UI', sans-serif; margin: 20px; background: #1a1a2e; color: #e0e0e0; }
-  h1 { color: #00d4ff; margin-bottom: 4px; }
-  h2 { color: #8892b0; margin-top: 24px; margin-bottom: 8px; font-size: 1.1em; }
-  .meta { color: #555; font-size: 0.82em; margin-bottom: 16px; }
-  .status-bar { display: flex; gap: 24px; padding: 12px 16px; background: #16213e; border-radius: 8px; margin-bottom: 20px; flex-wrap: wrap; }
-  .status-bar .item { display: flex; flex-direction: column; }
-  .status-bar .label { color: #556; font-size: 0.75em; text-transform: uppercase; letter-spacing: 0.5px; }
-  .status-bar .value { color: #e0e0e0; font-size: 1.1em; font-weight: 600; }
-  .status-bar .value.running { color: #00c853; }
-  .status-bar .value.idle { color: #8892b0; }
-  .status-bar .value.stopped { color: #ff5252; }
-  table { border-collapse: collapse; width: 100%; margin-bottom: 20px; }
-  th, td { text-align: left; padding: 6px 10px; border-bottom: 1px solid #2a2a4a; font-size: 0.9em; }
-  th { background: #16213e; color: #00d4ff; font-weight: 600; position: sticky; top: 0; }
-  tr:hover { background: #16213e; }
-  .complete { color: #00c853; }
-  .failed, .timeout { color: #ff5252; }
-  .never { color: #555; }
-  .lock_conflict, .incomplete { color: #ffa726; }
-  .disabled { opacity: 0.4; }
-  .evt-task_completed { color: #00c853; }
-  .evt-task_started { color: #42a5f5; }
-  .evt-task_skipped { color: #ffa726; }
-  .evt-scheduler_sleeping { color: #555; }
-  .evt-scheduler_started { color: #00d4ff; }
-  .pulse { display: inline-block; width: 8px; height: 8px; border-radius: 50%; background: #00c853; margin-right: 6px; animation: pulse 2s ease-in-out infinite; }
-  @keyframes pulse { 0%,100% { opacity: 1; } 50% { opacity: 0.3; } }
-</style>
-</head>
-<body>
-<h1>Scheduler Dashboard</h1>
-<p class="meta">Live &mdash; polls every 10s, instant updates via WebSocket</p>
-
-<div class="status-bar" id="status-bar">
-  <div class="item"><span class="label">Status</span><span class="value" id="sb-status">&mdash;</span></div>
-  <div class="item"><span class="label">Uptime</span><span class="value" id="sb-uptime">&mdash;</span></div>
-  <div class="item"><span class="label">Current</span><span class="value" id="sb-current">&mdash;</span></div>
-  <div class="item"><span class="label">Completed</span><span class="value" id="sb-completed">&mdash;</span></div>
-  <div class="item"><span class="label">Failed</span><span class="value" id="sb-failed">&mdash;</span></div>
-  <div class="item"><span class="label">Next</span><span class="value" id="sb-next">&mdash;</span></div>
-</div>
-
-<h2>Task Board</h2>
-<table>
-  <thead><tr><th>Task</th><th>Schedule</th><th>Last Run</th><th>Status</th><th>Duration</th><th>Overdue</th><th>Runs</th></tr></thead>
-  <tbody id="board-body"></tbody>
-</table>
-
-<h2>Recent Events</h2>
-<table>
-  <thead><tr><th style="width:70px">Time</th><th style="width:160px">Event</th><th style="width:180px">Task</th><th>Details</th></tr></thead>
-  <tbody id="events-body"></tbody>
-</table>
-
-<script>
-const BASE = location.origin;
-const WS_URL = BASE.replace(/^http/, "ws") + "/ws";
-
-function fmtInterval(s) {
-  s = Math.round(s);
-  if (s < 60) return s + "s";
-  if (s < 3600) return Math.floor(s / 60) + "m";
-  var h = Math.floor(s / 3600), m = Math.floor((s % 3600) / 60);
-  return m ? h + "h" + m + "m" : h + "h";
-}
-
-function fmtTime(iso) {
-  try { return new Date(iso).toLocaleTimeString([], {hour:"2-digit",minute:"2-digit",second:"2-digit"}); }
-  catch(e) { return iso ? iso.slice(11, 19) : "?"; }
-}
-
-function esc(s) {
-  var d = document.createElement("div");
-  d.textContent = s || "";
-  return d.innerHTML;
-}
-
-async function refreshLive() {
-  try {
-    var r = await fetch(BASE + "/api/scheduler/live");
-    if (!r.ok) throw new Error(r.status);
-    var d = await r.json();
-    var el = document.getElementById("sb-status");
-    if (d.current_task) { el.textContent = "running"; el.className = "value running"; }
-    else if (d.status === "running") { el.innerHTML = "<span class=\"pulse\"></span>idle"; el.className = "value idle"; }
-    else { el.textContent = d.status || "stopped"; el.className = "value stopped"; }
-    document.getElementById("sb-uptime").textContent = fmtInterval(d.uptime_seconds || 0);
-    if (d.current_task) {
-      var run = "";
-      if (d.current_task_started) {
-        var elapsed = (Date.now() - new Date(d.current_task_started).getTime()) / 1000;
-        run = " (" + fmtInterval(elapsed) + ")";
-      }
-      document.getElementById("sb-current").textContent = d.current_task + run;
-    } else { document.getElementById("sb-current").textContent = "\u2014"; }
-    document.getElementById("sb-completed").textContent = d.tasks_completed != null ? d.tasks_completed : "\u2014";
-    document.getElementById("sb-failed").textContent = d.tasks_failed != null ? d.tasks_failed : "\u2014";
-    if (d.next_task) {
-      document.getElementById("sb-next").textContent = d.next_task + " (in " + fmtInterval(d.next_in_seconds || 0) + ")";
-    } else { document.getElementById("sb-next").textContent = "\u2014"; }
-  } catch(e) {
-    document.getElementById("sb-status").textContent = "offline";
-    document.getElementById("sb-status").className = "value stopped";
-  }
-}
-
-async function refreshBoard() {
-  try {
-    var r = await fetch(BASE + "/api/scheduler/board");
-    if (!r.ok) return;
-    var d = await r.json();
-    var tbody = document.getElementById("board-body");
-    tbody.innerHTML = (d.tasks || []).map(function(t) {
-      var cls = t.enabled ? t.last_status : "disabled";
-      var label = esc(t.label) + (t.enabled ? "" : " <span style=\"color:#555\">[off]</span>");
-      return "<tr class=\"" + cls + "\">" +
-        "<td>" + label + "</td>" +
-        "<td>" + esc(t.schedule_str || "?") + "</td>" +
-        "<td>" + esc(t.last_run) + "</td>" +
-        "<td class=\"" + esc(t.last_status) + "\">" + esc(t.last_status) + "</td>" +
-        "<td>" + t.last_duration + "s</td>" +
-        "<td>" + esc(t.overdue_str) + "</td>" +
-        "<td>" + t.run_count + "</td></tr>";
-    }).join("");
-  } catch(e) {}
-}
-
-async function refreshEvents() {
-  try {
-    var r = await fetch(BASE + "/api/scheduler/events?tail=30");
-    if (!r.ok) return;
-    var d = await r.json();
-    var evts = (d.events || []).slice().reverse();
-    var tbody = document.getElementById("events-body");
-    tbody.innerHTML = evts.map(function(evt) {
-      var ts = fmtTime(evt.ts);
-      var etype = evt.event || "?";
-      var task = evt.task || "";
-      var detail = "";
-      if (etype === "task_completed") detail = esc(evt.status) + " \u2014 " + esc(evt.summary);
-      else if (etype === "task_skipped") detail = esc(evt.reason);
-      else if (etype === "scheduler_sleeping") detail = "next: " + esc(evt.next_task) + " in " + Math.round(evt.sleep_seconds || 0) + "s";
-      else if (etype === "task_started") detail = esc(evt.session);
-      else if (etype === "scheduler_started") detail = (evt.enabled_count || 0) + "/" + (evt.task_count || 0) + " tasks enabled";
-      return "<tr><td>" + ts + "</td><td class=\"evt-" + esc(etype) + "\">" + esc(etype) + "</td><td>" + esc(task) + "</td><td>" + detail + "</td></tr>";
-    }).join("");
-  } catch(e) {}
-}
-
-function connectWS() {
-  try {
-    var ws = new WebSocket(WS_URL);
-    ws.onmessage = function(e) {
-      try {
-        var msg = JSON.parse(e.data);
-        if (msg.type === "scheduler_update") { refreshLive(); refreshBoard(); refreshEvents(); }
-      } catch(ex) {}
-    };
-    ws.onclose = function() { setTimeout(connectWS, 5000); };
-    ws.onerror = function() { ws.close(); };
-  } catch(e) {}
-}
-
-refreshLive(); refreshBoard(); refreshEvents();
-setInterval(refreshLive, 10000);
-setInterval(refreshBoard, 10000);
-setInterval(refreshEvents, 10000);
-connectWS();
-</script>
-</body>
-</html>'''
-
-
 class VersionAction(argparse.Action):
     """Custom version action that checks Python version and pip environment."""
 
@@ -6154,74 +4294,6 @@ Command Categories:
     )
     mcp_parser.set_defaults(func=cmd_mcp)
 
-    # === ensure command (scheduled workloads) ===
-    ensure_parser = subparsers.add_parser(
-        "ensure",
-        help="Run named task with reliable session management",
-        description="Execute a task from .agentwire.yml with locking, retries, and completion detection.",
-    )
-    ensure_parser.add_argument("-s", "--session", required=True, help="Target session name")
-    ensure_parser.add_argument("-p", "--project", help="Project path containing .agentwire.yml (defaults to ~/projects/{session})")
-    ensure_parser.add_argument("--task", required=True, help="Task name from .agentwire.yml")
-    ensure_parser.add_argument("--dry-run", action="store_true", help="Show what would execute without running")
-    ensure_parser.add_argument("--wait-lock", action="store_true", help="Wait for lock instead of failing if locked")
-    ensure_parser.add_argument("--lock-timeout", type=int, default=60, help="Max time to wait for lock (default: 60s)")
-    ensure_parser.add_argument("--skip-if-locked", action="store_true", help="Exit 0 silently if session is locked (for cron use cases)")
-    ensure_parser.add_argument("--json", action="store_true", help="Output JSON")
-    ensure_parser.set_defaults(func=cmd_ensure)
-
-    # === task command group ===
-    task_parser = subparsers.add_parser(
-        "task",
-        help="Manage scheduled tasks",
-        description="List, show, and validate tasks defined in .agentwire.yml.",
-    )
-    task_subparsers = task_parser.add_subparsers(dest="task_command")
-
-    # task list
-    task_list = task_subparsers.add_parser("list", help="List tasks for session/project")
-    task_list.add_argument("session", nargs="?", help="Session name (default: current directory)")
-    task_list.add_argument("--json", action="store_true", help="Output JSON")
-    task_list.set_defaults(func=cmd_task_list)
-
-    # task show
-    task_show = task_subparsers.add_parser("show", help="Show task definition details")
-    task_show.add_argument("task", help="Task name (session/task or just task)")
-    task_show.add_argument("--json", action="store_true", help="Output JSON")
-    task_show.set_defaults(func=cmd_task_show)
-
-    # task validate
-    task_validate = task_subparsers.add_parser("validate", help="Validate task configuration")
-    task_validate.add_argument("task", help="Task name (session/task or just task)")
-    task_validate.add_argument("--json", action="store_true", help="Output JSON")
-    task_validate.set_defaults(func=cmd_task_validate)
-
-    # === lock command group ===
-    lock_parser = subparsers.add_parser(
-        "lock",
-        help="Manage session locks",
-        description="List, clean, and remove session locks.",
-    )
-    lock_subparsers = lock_parser.add_subparsers(dest="lock_command")
-    lock_parser.set_defaults(func=cmd_lock)
-
-    # lock list
-    lock_list_parser = lock_subparsers.add_parser("list", help="List all locks")
-    lock_list_parser.add_argument("--json", action="store_true", help="Output JSON")
-    lock_list_parser.set_defaults(func=cmd_lock_list)
-
-    # lock clean
-    lock_clean_parser = lock_subparsers.add_parser("clean", help="Remove stale locks")
-    lock_clean_parser.add_argument("--dry-run", action="store_true", help="Show what would be removed")
-    lock_clean_parser.add_argument("--json", action="store_true", help="Output JSON")
-    lock_clean_parser.set_defaults(func=cmd_lock_clean)
-
-    # lock remove
-    lock_remove_parser = lock_subparsers.add_parser("remove", help="Force-remove a lock")
-    lock_remove_parser.add_argument("session", help="Session name")
-    lock_remove_parser.add_argument("--json", action="store_true", help="Output JSON")
-    lock_remove_parser.set_defaults(func=cmd_lock_remove)
-
     # === scratchpad command group ===
     scratchpad_parser = subparsers.add_parser(
         "scratchpad",
@@ -6294,87 +4366,6 @@ Command Categories:
     services_down_parser.add_argument("--json", action="store_true", help="Output JSON")
     services_down_parser.set_defaults(func=cmd_services_down)
 
-    # === scheduler command group ===
-    scheduler_parser = subparsers.add_parser(
-        "scheduler",
-        help="Manage the task scheduler",
-        description=(
-            "Centralized daemon that dispatches tasks across projects on a shared cadence. "
-            "Tasks in ~/.agentwire/scheduler.yaml are either ensure tasks (task: + session:) "
-            "or workflow tasks (workflow: + inputs:) — the scheduler routes each automatically. "
-            "See docs/wiki/scheduling/scheduled-workloads.md."
-        ),
-    )
-    scheduler_subparsers = scheduler_parser.add_subparsers(dest="scheduler_command")
-
-    # scheduler start
-    sched_start = scheduler_subparsers.add_parser("start", help="Start scheduler daemon")
-    sched_start.set_defaults(func=cmd_scheduler_start)
-
-    # scheduler serve (foreground, for tmux)
-    sched_serve = scheduler_subparsers.add_parser("serve", help="Run scheduler in foreground")
-    sched_serve.set_defaults(func=cmd_scheduler_serve)
-
-    # scheduler stop
-    sched_stop = scheduler_subparsers.add_parser("stop", help="Stop scheduler")
-    sched_stop.set_defaults(func=cmd_scheduler_stop)
-
-    # scheduler status
-    sched_status = scheduler_subparsers.add_parser("status", help="Check scheduler status")
-    sched_status.add_argument("--json", action="store_true", help="Output JSON")
-    sched_status.set_defaults(func=cmd_scheduler_status)
-
-    # scheduler board
-    sched_board = scheduler_subparsers.add_parser("board", help="Show task board with overdue scores")
-    sched_board.add_argument("--json", action="store_true", help="Output JSON")
-    sched_board.set_defaults(func=cmd_scheduler_board)
-
-    # scheduler run <name>
-    sched_run = scheduler_subparsers.add_parser("run", help="Force-run a task now")
-    sched_run.add_argument("name", help="Task name from board")
-    sched_run.add_argument("--json", action="store_true", help="Output JSON")
-    sched_run.set_defaults(func=cmd_scheduler_run)
-
-    # scheduler enable <name>
-    sched_enable = scheduler_subparsers.add_parser("enable", help="Enable a task")
-    sched_enable.add_argument("name", help="Task name")
-    sched_enable.set_defaults(func=cmd_scheduler_enable)
-
-    # scheduler disable <name>
-    sched_disable = scheduler_subparsers.add_parser("disable", help="Disable a task")
-    sched_disable.add_argument("name", help="Task name")
-    sched_disable.set_defaults(func=cmd_scheduler_disable)
-
-    # scheduler history
-    sched_history = scheduler_subparsers.add_parser("history", help="Show recent run history")
-    sched_history.add_argument("--json", action="store_true", help="Output JSON")
-    sched_history.set_defaults(func=cmd_scheduler_history)
-
-    # scheduler events
-    sched_events = scheduler_subparsers.add_parser("events", help="Show recent scheduler events")
-    sched_events.add_argument("--json", action="store_true", help="Output JSON")
-    sched_events.add_argument("--tail", type=int, default=20, help="Number of events (default: 20)")
-    sched_events.add_argument("--task", help="Filter by task name")
-    sched_events.set_defaults(func=cmd_scheduler_events)
-
-    # scheduler live
-    sched_live = scheduler_subparsers.add_parser("live", help="Show live scheduler state")
-    sched_live.add_argument("--json", action="store_true", help="Output JSON")
-    sched_live.add_argument("--watch", action="store_true", help="Re-read every 2s")
-    sched_live.set_defaults(func=cmd_scheduler_live)
-
-    # scheduler dashboard
-    sched_dashboard = scheduler_subparsers.add_parser("dashboard", help="Open scheduler dashboard")
-    sched_dashboard.add_argument("--no-open", action="store_true", help="Generate HTML without opening")
-    sched_dashboard.set_defaults(func=cmd_scheduler_dashboard)
-
-    # scheduler report
-    sched_report = scheduler_subparsers.add_parser("report", help="Generate morning report of recent task runs")
-    sched_report.add_argument("--since", default="8h", metavar="DURATION", help="Time window (e.g. 8h, 12h, 1d) default: 8h")
-    sched_report.add_argument("--artifact", action="store_true", help="Open report as portal artifact")
-    sched_report.add_argument("--json", action="store_true", help="Output JSON")
-    sched_report.set_defaults(func=cmd_scheduler_report)
-
     # === Extracted command groups (each registrar owns its own subparser) ===
     # Phase 1 of #495 appends one entry here per extracted domain.
     #   - limits: usage-limit recovery (detect dialog, park, auto-resume)
@@ -6384,6 +4375,7 @@ Command Categories:
     from . import diff_cli, limits_cli, msg_cli, pane_cli, prompts_cli, send_cli  # noqa: I001  # session_cli kept on its own line below to minimize Phase 1 #495 merge conflicts
     from . import session_cli
     from . import channels_cli, portal_cli, tts_cli
+    from . import scheduler_cli, ensure_cli
     from .council import cli as council_cli
 
     _REGISTRARS = [  # noqa: N806  # registry constant; Phase 1 of #495 appends here
@@ -6397,6 +4389,8 @@ Command Categories:
         portal_cli.register_portal_parser,
         tts_cli.register_tts_parser,
         channels_cli.register_channels_parser,
+        scheduler_cli.register_scheduler_parser,
+        ensure_cli.register_ensure_parser,
     ]
     for _reg in _REGISTRARS:
         _reg(subparsers)
